refactors BP job methods into separate concerns
