+++ app/models/concerns/updatable.rb
@@ -299,5 +299,15 @@ module Updatable
       false
     end
   end
+
+  # CHECKS THAT METADATA SOURCE IS VALID - USED BY UPDATE
+  def validate_metadata_source(metadata_source, index)
+    if MetadataSource.all_metadata_cloud_names.include?(metadata_source)
+      true
+    else
+      batch_processing_event("Skipping row [#{index + 2}] with unknown metadata source: #{metadata_source}.  Accepted values are 'ladybird', 'aspace', 'sierra', or 'ils'.", 'Skipped Row')
+      false
+    end
+  end
 end
 # rubocop:enable Metrics/ModuleLength
